+++ gradle.properties
@@ -9,9 +9,9 @@ org.gradle.parallel=true
 	loader_version=0.14.19
 
 # Mod Properties
-	mod_version = 1.0.2
-	maven_group = com.rebane2001
-	archives_base_name = aimobs
+	mod_version = 0.0.1
+	maven_group = com.jackdaw
+	archives_base_name = chat-with-npc
 
 # Dependencies
 	fabric_version=0.80.0+1.19.4
