+++ src/main/resources/chat-with-npc.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "com.rebane2001.aimobs.mixin",
+  "package": "com.jackdaw.chatWithNPC.mixin",
   "compatibilityLevel": "JAVA_17",
   "client": [
     "ChatHudAccessor",
